+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [unreleased]
 
+### Fixed
+
+- Fixed `/compute/output/{task_id}` endpoint to properly return `*Output` or `ProgramFailure` objects for each calculation in a list of inputs.
+
+### Removed
+
+- Removed distinction between `Result` and `ResultGroup` and created a single output object for the `/compute/output/{task_id}` endpoint, `Output`.
+
 ## [0.11.0] - 2023-09-08
 
 ### Added